+++ CHANGELOG.md
@@ -19,6 +19,7 @@ All notable changes to this project will be logged here
 - ✏️ move store selection to `useActiveStore.ts`
 - ✏️ add a version check to prevent build with same version number
 - ✏️ remove nested divs in `About.tsx`
+- ✏️ move scroll to top button to `ScrollToTop.tsx` and added to `About.tsx` and `WeeklyTracker.tsx` pages
 
 ## 0.1.10